test: move TriggerButton style test
